Newbranch (#4)
* random bs go

* Initialize project

* Add Card designs

* Flip Cards and Game bar

Added a side bar to track the games progess.
Added js to track if the crads match up. If not, unflip them.

* Designs + Reset

Addded more custom designs for the cards.

Added a reset button for the game

* Attempting to add players

An so far fruitless attempt on adding difficulty and players

* Trying to fix up two player mode

Attemplting to get the playerscore properly displayed and reset.

* Tring to get the reset game function to work

* Revert "Tring to get the reset game function to work"

This reverts commit 63f884df7d56e9f8e9861e32225e800a3ce88494.

* Fixed two player display + Timer

Fixed the Match number display for Player two.
Added a Timer that starts Running as soon as the page is loaded.

* Track if all matches found

Added a function to track when all cards are  flipped.
Timer gets stopped when this is the case.

* Improved restart Button

The restart button now executes several functions that reset all aspects of the game

* Vanish cards when found

Added a fuction to make the cards backface dissaper so they aren't visible anymore. Also added a fuction to revet the change tied to the reset button.

* Added Background Ambience

Added backround audio, that can be toggeled via button.

* Sound effects

Added sound effects for flipping, matching and resset.

* changed font

* startedd styling

Started styling the textareas

* Almost done

Styled the entire thing

* Finalizing

Swapped Background with a hand drawn one.

Final trouble shooting

* Final

Commenting the code
and clearing it out a little.